requirements.txt
Django==4.2.7
djangorestframework==3.14.0
rest_framework_simplejwt
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
django-filter==23.1
psycopg2-binary>=2.9.0

# New notification dependencies
celery>=5.2.0
redis>=4.3.0
twilio>=8.2.0
sendgrid>=6.9.0
